requirements.txt
python-nmap
requests
beautifulsoup4
reportlab
slack_sdk

